[FLINK-32859][scheduler] Introduce the StateWithoutExecutionGraph
